finish 4.15 - 4.23

+++ controllers/blog.js
@@ -1,29 +1,50 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response, next) => {
     try {
-      const blogs = await Blog.find({})
+      const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
       response.json(blogs)
     } catch (error) {
       next(error)
     }
 })
   
-blogRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+blogRouter.post('/', userExtractor, async (request, response, next) => {
+  const body = request.body
+  const user = request.user
+
+  const blog = new Blog({
+    ...body,
+    user: user._id
+  })
 
   try {
-    const result = await blog.save()
-    response.status(201).json(result)
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.status(201).json(savedBlog)
   } catch (error) {
     next(error)
   }
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === user._id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id )
+    await user.save()
+
+    return response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'Unauthorised resource' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
